Add responsive navbar with user interaction
A responsive navigation bar was integrated into the website. In the form of a CSS and JavaScript combination, it allows for seamless user interaction on both mobile and desktop devices. This commit also includes various updates in the login/signup system, including a proper log out button, along with other necessary navigation updates.