Add numeric validation to various APIs
Added numeric validation checks to prevent potential type errors in various API endpoints including edit-user, authentication, points-helper, and modify-teams. Numeric validation helps ensure the received data is of the correct type, increasing the reliability and stability of the system. This also includes a minor fix to cleanse the hex code for color in the modify-teams API.