Resolve #9
Added the ability to create, edit, and view descriptions when adding points and viewing them from the history tab. This commit also adds an edited flag to the descriptions when a points entry is edited. Additionally, the rounded corners on the top of the history page were not working, so that was fixed. The scroll bar was also removed from the history tab as that messed up the view width as well.