Add intermediary api files to make requests work
In order to get requests to files sitting behind the html folder, you need to make them from the machine. To do this, new files have been created which act in between the user and these api files. To access them, the user must supply a correct auth string for verification, and only then can the api files get called.